+++ tests/test_shardformer/test_model/test_shard_gpt2.py
@@ -106,9 +106,17 @@ def check_forward_backward(model_fn, data_gen_fn, output_transform_fn, loss_fn, 
     'use_lazy_init': False,
     'precision': 'fp32',
 }, {
+    'tp_size': 2,
+    'pp_size': 2,
+    'num_microbatches': 4,
+    'enable_all_optimization': True,
+    'use_lazy_init': True,
+    'enable_sequence_parallelism': True,
+    'precision': 'fp32',
+}, {
     'tp_size': 4,
     'pp_size': 1,
-    'enable_all_optimization': False,
+    'enable_all_optimization': True,
     'use_lazy_init': True,
     'enable_sequence_parallelism': True,
     'precision': 'fp32',
